Move svg-to-react conversion out of the source-specific code
